Make the code more OOP - Refactored code by separating them to folders ui, task, storage and parser
