+++ src/main/java/seedu/shao/task/Deadline.java
@@ -1,13 +1,17 @@
-package seedu.shao;
+package seedu.shao.task;
 
 import java.time.LocalDateTime;
 
 public class Deadline extends Task {
-	protected LocalDateTime by;
+	private LocalDateTime by;
 
 	public Deadline(String description, LocalDateTime by) {
 		super(description);
 		this.by = by;
+	}
+
+	public LocalDateTime getBy() {
+		return by;
 	}
 
 	@Override
