+++ src/main/java/seedu/shao/task/Task.java
@@ -1,4 +1,4 @@
-package seedu.shao;
+package seedu.shao.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
