+++ src/test/java/seedu/shao/DeadlineTest.java
@@ -1,6 +1,9 @@
 package seedu.shao;
 
 import org.junit.jupiter.api.Test;
+
+import seedu.shao.task.Deadline;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.time.LocalDate;
